feat : DDL 인터페이스 구조 변경
