+++ src/main/java/persistence/sql/Dialect.java
@@ -1,0 +1,7 @@
+package persistence.sql;
+
+import jakarta.persistence.GenerationType;
+
+public interface Dialect {
+    String getGenerationTypeQuery(GenerationType generationType);
+}
