+++ src/main/java/persistence/sql/ddl/CreateQueryBuilder.java
@@ -1,5 +1,54 @@
 package persistence.sql.ddl;
 
-public interface CreateQueryBuilder {
-    String makeQuery();
+import jakarta.persistence.Entity;
+import persistence.sql.Dialect;
+import persistence.sql.model.TableName;
+
+public class CreateQueryBuilder implements QueryBuilder {
+    private static final String LEFT_PARENTHESIS = "(";
+    private static final String RIGHT_PARENTHESIS = ")";
+    private static final String SPACE = " ";
+
+    private final Class<?> clazz;
+    private final Dialect dialect;
+
+    public CreateQueryBuilder(Class<?> clazz, Dialect dialect) {
+        ExceptionUtil.requireNonNull(clazz);
+
+        if (!clazz.isAnnotationPresent(Entity.class)) {
+            throw new IllegalArgumentException("Entity 클래스가 아닙니다.");
+        }
+
+        this.clazz = clazz;
+        this.dialect = dialect;
+    }
+
+    @Override
+    public String build() {
+        StringBuilder makeStringBuilder = new StringBuilder();
+        makeStringBuilder.append(createTableIfNotExistsStatement());
+        makeStringBuilder.append(generateColumnDefinitions());
+        return makeStringBuilder.toString();
+    }
+
+    private String generateColumnDefinitions() {
+        DDLColumn ddlColumns = new DDLColumn(this.clazz.getDeclaredFields(), dialect);
+
+        StringBuilder columnDefinitionStringBuilder = new StringBuilder();
+        columnDefinitionStringBuilder.append(LEFT_PARENTHESIS);
+        columnDefinitionStringBuilder.append(ddlColumns.makeColumnsDDL());
+        columnDefinitionStringBuilder.append(RIGHT_PARENTHESIS);
+        return columnDefinitionStringBuilder.toString();
+    }
+
+    private String createTableIfNotExistsStatement() {
+        TableName tableName = new TableName(clazz);
+
+        StringBuilder stringBuilder = new StringBuilder();
+        stringBuilder.append("CREATE TABLE IF NOT EXISTS");
+        stringBuilder.append(SPACE);
+        stringBuilder.append(tableName.getValue());
+
+        return "CREATE TABLE IF NOT EXISTS " + tableName.getValue();
+    }
 }
