+++ src/main/java/persistence/sql/ddl/DropQueryBuilder.java
@@ -1,5 +1,34 @@
 package persistence.sql.ddl;
 
-public interface DropQueryBuilder {
-    String makeQuery();
+import jakarta.persistence.Entity;
+import persistence.sql.model.TableName;
+
+public class DropQueryBuilder implements QueryBuilder {
+
+    private static final String SPACE = " ";
+
+    private static final String DROP_TABLE_DDL = "DROP TABLE IF EXISTS";
+
+
+    private final Class<?> clazz;
+
+    public DropQueryBuilder(Class<?> clazz) {
+        ExceptionUtil.requireNonNull(clazz);
+
+        if (!clazz.isAnnotationPresent(Entity.class)) {
+            throw new IllegalArgumentException("Entity 클래스가 아닙니다.");
+        }
+        this.clazz = clazz;
+    }
+
+    @Override
+    public String build() {
+        StringBuilder dropTableQueryStringBuilder = new StringBuilder();
+
+        TableName tableName = new TableName(this.clazz);
+        dropTableQueryStringBuilder.append(DROP_TABLE_DDL);
+        dropTableQueryStringBuilder.append(SPACE);
+        dropTableQueryStringBuilder.append(tableName.getValue());
+        return dropTableQueryStringBuilder.toString();
+    }
 }
