+++ src/main/java/persistence/sql/ddl/QueryBuilder.java
@@ -1,0 +1,7 @@
+package persistence.sql.ddl;
+
+public interface QueryBuilder {
+
+    String build();
+
+}
